Airflow/requirements.txt
apache-airflow-providers-google
apache-airflow[postgres]
pyarrow
psycopg2-binary
python-dotenv 
